refactor(discord): rename necord-options to necord.config
