+++ apps/scrubjay-discord/src/discord/necord.config.ts
